test: add new tests for JITCHomoR event-based operations in _jitc_homo_test.py
